Reorder and group the related commands together
